create driver hook
