+++ src/index.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import "driver.js/dist/driver.css";
 import "./index.css";
 import App from "./App";
 // @ts-ignore
