+++ tailwind.config.js
@@ -3,7 +3,10 @@
 const withMT = require("@material-tailwind/react/utils/withMT");
 
 module.exports = withMT({
-  content: ["./src/**/*.{js,jsx,ts,tsx}"],
+  content: [
+    "./src/**/*.{js,jsx,ts,tsx}",
+    "./node_modules/driver.js/dist/**/*.js",
+  ],
   theme: {
     fontSize: {
       xs: [
